chore: refactor navigation menus to use constants
Refactor the code in `layouts/default.vue` and `constants/menus.ts` to use constants for the main and secondary menus. This improves code organization and maintainability. The `mainMenus` and `secondaryMenus` constants are now imported from `~/constants/menus` instead of being defined locally. This change ensures consistency and makes it easier to update the menus in the future.